vts: allow no endorsements inside GetAttestation
A prior assumption was that there would always be endorsements
associated with a token. As a result failing to retrieve any
endorsements always resulted in an error.

In retrospect, that assumption may not hold for all possible attestation
schemes. Update GetAttestation implementation to allow the situation
where no endorsements are associated with a token by only propagating
the returned error if it is not kvstore.KeyNotFound.

Signed-off-by: Sergei Trofimov <sergei.trofimov@arm.com>