kvstore: add Setup() method
Add IKVStore.Setup() that performs one-time setup of the store. This is
a  no-op for memory backend. For sql, this creates the table used by the
store. This is intended to be used by CLI tools to remove the need for
the manual setup on half of the user during deployment.

Signed-off-by: Sergei Trofimov <sergei.trofimov@arm.com>