vts: allow listening on multiple interfaces.
Previously, VTS server used `vts.server-addr` configuration as the
listening address. This, however, restricted to listening on a single
network interface. Now, the server only uses the port form the setting,
and listens on all system interfaces. For edge cases where restricting
the interface is desired, `vts.listen-addr` setting is introduced which
overrides `server-addr` if specified.

Signed-off-by: Sergei Trofimov <sergei.trofimov@arm.com>