Remove AttestationFormat
We currently have three was of referencing attestation scheme specific
functionality:

- media type
- attestation format
- attestation scheme name (implemented as string version of the above)

This change removes attestation format because:

- It is currently not being set properly (end-to-end logs show it as
  unset during verification path execution).
- It is not consistently used (which is why the above has gone unnoticed
  until now).
- It is made redundant by the media type and scheme name
- The enum defining valid values is hard-coded, meaning adding new
  plugins requires a modification to the core services code line.

In the situations where attestation format being used, the scheme name
is used instead.

Note: currently, this means that the scheme name must be hard-coded in
two places -- the scheme plugin, and the provisioning decoder plugin;
and those two values must match. The upcoming plugin framework
unification and restructuring will remove dependency and allow the name
to be defined in a single place from which it will be used by both
plugins.

Signed-off-by: Sergei Trofimov <sergei.trofimov@arm.com>