end-to-end: keep tmux pane after service exit
Ensure that the tmux pane in which a server was running remains after
the service exits. This allows going back in the output history to see
what went wrong. This also results in less confusing behavior if
vts-services fails to start (the old behavior resulted in the entire
session being immediately terminated, and the subsequent commands to
start other services giving weird messages that session "veraison" did
not exist).

Signed-off-by: Sergei Trofimov <sergei.trofimov@arm.com>