Add debug config files
Add config.yaml to the provisioning and verification services
directories that assumes they are being executed from a debug shell on
the docker network. This saves having to re-create them every time one
needs to debug the service.

Note: we should allow specifying the location of the config via a
command line option, which will remove the need to keep these files in
the source (we could then just point them to the deployment config when
debugging).

Signed-off-by: Sergei Trofimov <sergei.trofimov@arm.com>