Handle evidence errors
Handle evidence errors in a consistent manner that is different from
other errors. This is done via a new error type, BadEvidenceError, which
schemes and VTS return to indicate a problem with the provided evidence,
and not a function within the service.

Evidence errors are not propagated by the REST API. Instead, they are
reported as part of the attestation result (All AR4SI claims set to
value 1).

Signed-off-by: Sergei Trofimov <sergei.trofimov@arm.com>