integ-test: add TEST_USER env variable
Allow specifying of the user the debug shell will be run as via
TEST_USER environment variable (similar to how DEBUG_USER is used in the
docker deployment Makefile).

This allows to start the test runner container as root in order to make
changes to it prior to su'ing into the tavern user in order to run the
tests.

Signed-off-by: Sergei Trofimov <sergei.trofimov@arm.com>