integ-test: additional utils for tavern container
Add the cbor-diag utilities and step tool to the taster container. The
former allows examining the provisioning and evidence tokens (that are
encoded as CBOR), the latter allows examining attestation results
(encoded as JWT).

Also add vim, as a text editor is useful for quickly examining/adjusting
things from inside the debug shell.

Signed-off-by: Sergei Trofimov <sergei.trofimov@arm.com>