integ-test: rename generated endorsements
Rename the generated endorsements CoRIMs to follow the pattern
"corim-{scheme}-{corim-name}.cbor". Previously the generated file was
always called "endorsements.cbor", so endorsements for different test
cases were overwriting each other. This was not a problem for tests, as
each test regenerated the endorsements every time. However, this made it
difficult to get endorsements for a specific test (e.g. when debugging
or to be used in another context). One would need to re-run just that
one test; and afterwards, there would always be some confusion as to
what a particular "endorsements.cbor" contains.

Signed-off-by: Sergei Trofimov <sergei.trofimov@arm.com>