Consistent AppraisalPolicyID and scheme-specific policies
"Policy" as defined by the RATS architecture[1], and, therefore, the
ear.appraisal-policy-id entry n EAR[2], maps onto the combination of
attestation scheme and policy in Veraison.

This means that, when a policy is not used, the ear.appraisal-policy-id
field should be set to reflect the attestation scheme. If a policy is
used, the field should be set to reflect both, the scheme and the
policy.

Additionally, up to this point, the policy manager allowed only one
active policy per tenant. Differentiation between schemes, if necessary,
could be performed within the policy rules. This commit changes this so
that the polices are now managed based on both, the tenant and the
scheme. This means that policies for different schemes can be updated
independently by the tenant.

[1]: https://datatracker.ietf.org/doc/html/draft-ietf-rats-architecture-05
[2]: https://www.rfc-editor.org/rfc/internet-drafts/draft-fv-rats-ear-00.html

Signed-off-by: Sergei Trofimov <sergei.trofimov@arm.com>