Expose nonce to policies
The session nonce is intended to ensure freshness. It is intended to be
used as a challenge to the attester to ensure a fresh attestation report
is generated. It is also included in the attestation result. Thus far,
however, do not process it. The assumption is that this is done by the
RP. However, given that this is fairly generic, the RP may wish to rely
on the verifier for this. Given that a scheme may not be doing this
(none of our current implementations do), the policy would be a natural
place to do this on per-deployment basis.

Prior to this, the nonce established during the creation of a
verification session is added to the attestation result by the core
verification pipeline code, and is also available to scheme handlers
inside ExtractClaims(). However, the nonce was not provided to policies.

As there is no obvious place for the nonce within existing policy engine
inputs, this change adds a new input for this purpose. It also
implements a couple of fixes to the policy code.

- Amend policy agent interface to take an "appraisal context" that
  currently only contains the verification session's nonce. In the
  future, it may be used for any other items that need to be passed to
  the policy and do not fall within the existing categories of
  "evidence, "endorsements", or "results".
- Fix how evidence is expected to look to the agent. Previously, due to
  a mistake in the unit test inputs, the evidence was expected to be
  under the key "evidence" (where as the entire input map should be
  treated as evidence).
- Update github.com/veraison/ear dependency to v1.1.2. This adds updated
  wider limits for nonce sizes allowed inside eat (this is needed as CCA
  nonce exceeds the old limit when base64-encoded).
- Update CCA_SSD_PLATFORM scheme handler to extract both platform and
  realm claims into evidence. The scheme only processes platform
  evidence, so realm claims were not being extracted; however the
  session nonce is contained within the realm claims. The scheme handler
  still only validates the platform claims, however the realm claims
  (and the nonce) are now available to policies.
- Add an integration test case to show how a policy could be used to
  verify the nonce for the CCA_SSD_PLATFORM scheme.

Signed-off-by: Sergei Trofimov <sergei.trofimov@arm.com>