verification: nonce size validation on session create
Up to this point, on session creation, nonceSize parameter was required
to be a non-zero uint8 (i.e. in range 1..255), and the nonce parameter
was only verified to be  a valid base64 string, with no size limit.

The nonce eventually ends up inside EAR's eat_nonce claim, and therefore
must comport with the restriction associated with that claim, which is
between 8 and 64 bytes long. This will result in a failed attestation
result generation in the subsequent submit request.

This commit modifies nonceSize limit check to be aligned with eat_nonce
definition, and adds an analogous check for the nonce parameter. This
will insure that attempting to define a nonce of an invalid size during
session creation will result in a failure to create the session.

Signed-off-by: Sergei Trofimov <sergei.trofimov@arm.com>