Remove pre hook from scheme/common/Makefile
This hook does not appear to serve a purpose, and it triggers tests
during docker deployment creation (by invoking "make -C ../../handler",
which triggers the default target, which is "test"). This, in turn,
causes copyright checks to run as part of docker creation which can
cause problems in CI as these checks would not be property configured.

Signed-off-by: Sergei Trofimov <sergei.trofimov@arm.com>