feat(docker): add gen-certs script
Add a script to (re-)generate all certificates necessary for a Veraison
services deployment. The script makes some assumptions about what the
deployment looks like (e.g. how the service nodes are named) that is
based on the docker deployment, but should be generic enough to be
useful for other deployments.

The services certs are signed with a cert named "rootCA.crt" in the same
directory as the script (and the certs that will be generated). The
script will generate a self-signed rootCA.crt if one isn't provided.

Signed-off-by: Sergei Trofimov <sergei.trofimov@arm.com>