fix(auth/keycloak): clear flags before parsing
Keycloak's logging lib expects flags to be parsed, otherwise it
complains, so we're parsing them. The problem is that if we actually
have any flags specified, parsing will fail, as the lib doesn't know
about Veraison flags, so clear os.Args of flags before calling
flags.Parse().

This is fine as KeycloakAuthorizer.Init(), where the call occurs, is
invoked well after the service command line has already been processed,
so os.Args won't be ever be used from then on.

This does mean that we can no longer pass configuration for Keycloak
logging lib on the command line, which was theoretically possible
before. However we've literally never do this, it's completely
undocumented, and may potentially cause conflicts with Veraison's own
logging. So this fixes that as well.

Signed-off-by: Sergei Trofimov <sergei.trofimov@arm.com>