chore: update to latest CoRIM implementation
Update to support the latest github.com/veraison/corim, which is (more)
aligned with he latest CoRIM spec draft. The latest implementation has
the following major changes:

- UnsignedCorim's Profiles field (a slice of strings) has been replaced
  with Profile field (a single string).
- Reference value can now only contain a single measurement, so multiple
  measurements require multiple reference values (each with its own
  environment).
- CBOR keys for DevIdentityKeys and AttestVerifKeys have been swapped
  to align with the spec (this was a bug in the old implementation).

As the latest github.com/veraison/corim requires Go 1.22, the required
version for this code base has been bumped to that as well.

Signed-off-by: Sergei Trofimov <sergei.trofimov@arm.com>