feat(kvstore): add support for MySQL/MariaDB and Postgres
Add support for "mysql" and "pgx" drivers (for MySQL/MariaDB and
Postgres, respectively). This required the following changes to the
"sql" kvstore backend:

- "key" column had to be renamed to "kv_key", as the former is a
  reserved word in MySQL. "vals" column has also been renamed to
  "kv_val" for consistency (also, the old name implied that the column
  contained an array of values, where as it is a single string value).
- Queries are now assembled using squirrel, which allows switching the
  placeholder symbol. This is necessary because MySQL and Postgres
  expect different syntax for placeholders.

Documentation for KVStore has been amended to reflect that only three
drivers can be enabled via configuration for the services, and the
described support for other drivers is in the context of other code
using the KVStore implementation. (Plus minor line length
re-formatting.)

Native deployment has been augmented with optional config/init scripts
for the other SQL drivers, and documentation on how to enable them.

Signed-off-by: Sergei Trofimov <sergei.trofimov@arm.com>