chore: remove support for running services in-place
This is now possible via the native deployment (by deploying with
symlinks), and this just adds an unnecessary maintenance burden.

Note: this basically just adds a note saying that we don't support it,
and points to the native deployment. The ability to do so is still
there. However, if you're exercising it, you're on your own.

Signed-off-by: Sergei Trofimov <sergei.trofimov@arm.com>