fix!(scheme, api): correctly handle media type params with absolute URIs
When the "profile" parameter value is an absolute URI, it must be quoted.

The definition of parameter-value in Appendix A of RFC9110 allows both token
and quoted-string, but only the latter is compatible with absolute URI syntax,
which includes at least the (reserved) ':' following the scheme name.

(Note that the OID form for the profile does not need to be quoted.)

This change also makes sure that both the media types we read and those we
write are validated and normalised.

Signed-off-by: Thomas Fossati <thomas.fossati@linaro.org>