feat(native): deploy into intermediate DEPLOYMENT_DEST
Add DEPLOYMENT_DEST configuration variable that specifies the location
where the deployment will be generated. This defaults to VERAISON_ROOT
(the install location of the deployment), but may be set to a different
value, allowing the deployment to be created in an intermediate location
before being moved into its final installation destination.

This is useful when the native deployment is intended to be used as an
input for an alternative deployment mechanism (e.g. a Debian package)
rather than being used directly.

Signed-off-by: Sergei Trofimov <sergei.trofimov@arm.com>