chore: include version from git tag in builds
Up to now, the version reported by Veraison services was set during
build to "commit-$commit" where $commit is the short hash of the HEAD
commit during the build.

Since 9509e742, we now have regular auto-tagging on the repo in the form
"v0.0.yyMM". This should be included in the version.

This commit adds scrips/get-veraison-version that generates a version of
the form "$version+$commit" where $version is the version from tag sans
the leading "v". It also allows overriding this version using
VERAISON_BUILD_VERSION environment variable. The override may be useful
when git is not available (e.g. when building form a tarball) or when
wanting to encode additional information in the version (e.g. when
creating a build for a specific demo).

This script is then used to populate VERSION_FROM_GIT inside
mk/common.mk which is then compiled into the built executables.

Signed-off-by: Sergei Trofimov <sergei.trofimov@arm.com>