feat(kvstore): add max_connections setting for sql backend
Add max_connections setting for the sql backend that governs the maximum
number of parallel connections that will be opened by the store. Once
this limit is reached, further queries will block until a connection
will become available.

note: since VTS has three stores, the total maximum is three times this
value (assuming its the same for all stores).

The default is set to 10 (so 30 for Veraison as a whole) for most cases.
The server default seems to be 100 for both MySQL/MariaDB and
PostgreSQL; and this way Veraison shouldn't adversely affect performance
for other services that may be using the DBMS.

In AWS deployment it is exposed as a configuration and is set to 8.
Which is 72 in total (8 x 3 stores x max 3 instances/store). This
deployment creates a dedicated RDS instance, so other users are not a
concern. The max connections for RDS is determined dynamically, and
seems to be around 80 for the t3.micro (the default RDS instance type
used by the deployment). With total 75 connections, this leaves a little
headroom (it should also be noted that 3 connection slots are reserved
for superusers and are generally unavailable).

Signed-off-by: Sergei Trofimov <sergei.trofimov@arm.com>