fix(aws): use memcached for sessions
Switch to using memcached backend for session manager in AWS deployment.

- Rename "rds" stack to "support" to reflect the more general nature of
  the stack. Up to this point, it has solely been used for the RDS
  instances (and the sentinel which is needed for initializing the RDS
  instance). However the intent is to use it for anything that needs to
  exist prior to deploying services (e.g. because services'
  configuration needs to reflect it).
- Add an ElastiCache instance to the support stack. This is the AWS'
  caching services (kinda like what RDS is to MySQL/Postgres, this is to
  reddis/memcached).
- Change services configuration to use the ElastiCache instance for
  session management.

This fixes a problem when there is more than one verification instance
active behind the load balancer. It is possible (and, in fact, likely)
that the session creation request and the subsequent attestation request
go to different instances. When using local session cache, the second
instance won't have the session created on the first, resulting in an
error.

Signed-off-by: Sergei Trofimov <sergei.trofimov@arm.com>