fix(aws): use .well-known for health checks
Switch to using .well-known API for instance health checks.

Prior to this, root path with used, which confirmed that the HTTP
server was up and listening. However, VTS can take longer to start than
the HTTP server (due to plugin load?), so sending requests immediately
may result in error. .well-known API depends on VTS, so once that is
successful, we know VTS is up, and the service is actually ready.

AWS load balancer uses heath checks when choosing which instance to
route to, so by switching the health check to .well-known, we avoid
temporary failures when a new instance is brought up by auto-scaling.

Signed-off-by: Sergei Trofimov <sergei.trofimov@arm.com>