fix(aws): create linked role prior to support stack
The service linked role for ElastiCache can only be created once
per-account, and there does not appear to be a good way to do so
conditionally inside the template, while also maintaining correct
ordering.

So, instead of specifying it in the template, check if it exists and
create it if necessary before the stack.

Signed-off-by: Sergei Trofimov <sergei.trofimov@arm.com>