fix(deployments): ensure systemd units for deb and rpm
Add -S flag to native deployment to force creation of systemd units in
the deployment, regardless of the native system (there is also an
equivalent -L for launchd). Use -S inside dep and rpm deployments to
ensure that the packages are always created with systemd units even if
the local system doesn't have systemd.

Signed-off-by: Sergei Trofimov <sergei.trofimov@arm.com>