fix(aws): do not create boto3 session if not necessary
In boto3, clients for specific AWS services can be created by first
creating a session, or directly from the module, in which case a default
session initialized by the module will be used.

It seems that when creating a session, that session will not always pick
up AWS credentials (empirically, it seems to pick up config from
~/.aws/, but not from the environment variables). The default session
seems to check all sources.

To allow maximum flexibility for how credentials may be configured, do
not create a Session unless custom parameters are specified (which, in
the current implementation, they never are), and rely on the default
when creating clients.

Signed-off-by: Sergei Trofimov <sergei.trofimov@arm.com>