fix(deployments/rpm): build, install and package with RPM
RPM allows the ability to build, install and package the project. Add these
support for the server RPM. As a result, we don't need Makefile build support
for RPMs.

note: Since the RPMs are for linux-like distros, we aren't packaging "launchd"
files, as they're Darwin-specific. Also, we don't need to package "env" files,
as we already install the executables in the PATH.

Signed-off-by: Jagannathan Raman <jag.raman@oracle.com>